requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
pydantic
requests
transformers
sentence-transformers
pinecone-client
scikit-learn
pandas
scipy
